+++ base/client.jl
@@ -4,16 +4,28 @@
 ##             and REPL
 
 const text_colors = AnyDict(
-    :black   => "\033[1m\033[30m",
-    :red     => "\033[1m\033[31m",
-    :green   => "\033[1m\033[32m",
-    :yellow  => "\033[1m\033[33m",
-    :blue    => "\033[1m\033[34m",
-    :magenta => "\033[1m\033[35m",
-    :cyan    => "\033[1m\033[36m",
-    :white   => "\033[1m\033[37m",
-    :normal  => "\033[0m",
-    :bold    => "\033[1m",
+    :black         => "\033[30m",
+    :red           => "\033[31m",
+    :green         => "\033[32m",
+    :yellow        => "\033[33m",
+    :blue          => "\033[34m",
+    :magenta       => "\033[35m",
+    :cyan          => "\033[36m",
+    :white         => "\033[37m",
+    :dark_gray     => "\033[90m",
+    :light_red     => "\033[91m",
+    :light_green   => "\033[92m",
+    :light_yellow  => "\033[93m",
+    :light_blue    => "\033[94m",
+    :light_magenta => "\033[95m",
+    :light_cyan    => "\033[96m",
+    :normal        => "\033[0m",
+    :default       => "\033[39m",
+    :bold          => "\033[1m",
+)
+
+const disable_text_style = AnyDict(
+    :bold => "\033[22m",
 )
 
 for i in 0:255
@@ -22,7 +34,7 @@ end
 
 # Create a docstring with an automatically generated list
 # of colors.
-const possible_formatting_symbols = [:normal, :bold]
+const possible_formatting_symbols = [:normal, :bold, :default]
 available_text_colors = collect(filter(x -> !isa(x, Integer), keys(text_colors)))
 available_text_colors = cat(1,
     sort(intersect(available_text_colors, possible_formatting_symbols), rev=true),
@@ -35,11 +47,14 @@ const available_text_colors_docstring =
 """Dictionary of color codes for the terminal.
 
 Available colors are: $available_text_colors_docstring as well as the integers 0 to 255 inclusive.
+
+The color `:default` will print text in the default color while the color `:normal`
+will print text with all properties reset to normal (including boldness, italics etc.).
 """
 text_colors
 
 have_color = false
-default_color_warn = :red
+default_color_warn = :light_red
 default_color_info = :cyan
 if is_windows()
     default_color_input = :normal
@@ -59,8 +74,10 @@ end
 
 warn_color()   = repl_color("JULIA_WARN_COLOR", default_color_warn)
 info_color()   = repl_color("JULIA_INFO_COLOR", default_color_info)
-input_color()  = text_colors[repl_color("JULIA_INPUT_COLOR", default_color_input)]
-answer_color() = text_colors[repl_color("JULIA_ANSWER_COLOR", default_color_answer)]
+
+# Print input and answer in bold unless the user specifies :default in the ENV var.
+input_color()  = text_colors[:bold] * text_colors[repl_color("JULIA_INPUT_COLOR", default_color_input)]
+answer_color() = text_colors[:bold] * text_colors[repl_color("JULIA_ANSWER_COLOR", default_color_answer)]
 
 function repl_cmd(cmd, out)
     shell = shell_split(get(ENV,"JULIA_SHELL",get(ENV,"SHELL","/bin/sh")))
@@ -101,8 +118,10 @@ end
 
 display_error(er) = display_error(er, [])
 function display_error(er, bt)
-    with_output_color(:red, STDERR) do io
-        print(io, "ERROR: ")
+    with_output_color(:light_red, STDERR) do io
+        with_output_color(:bold, STDERR) do io_bold
+            print(io_bold, "ERROR: ")
+        end
         showerror(io, er, bt)
         println(io)
     end